requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
APScheduler==3.11.0
asyncer==0.0.8
attrs==25.1.0
audioop-lts==0.2.2
backoff==2.2.1
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
chainlit==2.7.1.1
charset-normalizer==3.4.2
chevron==0.14.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
Deprecated==1.2.18
discord.py==2.6.2
distro==1.9.0
fastapi==0.116.1
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.5
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Lazify==0.4.0
literalai==0.1.201
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.13.1
mdurl==0.1.2
monotonic==1.6
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==1.27.1
nest-asyncio==1.6.0
numpy==2.2.3
openai==1.102.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-alephalpha==0.45.6
opentelemetry-instrumentation-anthropic==0.45.6
opentelemetry-instrumentation-bedrock==0.45.6
opentelemetry-instrumentation-chromadb==0.45.6
opentelemetry-instrumentation-cohere==0.45.6
opentelemetry-instrumentation-crewai==0.45.6
opentelemetry-instrumentation-google-generativeai==0.45.6
opentelemetry-instrumentation-groq==0.45.6
opentelemetry-instrumentation-haystack==0.45.6
opentelemetry-instrumentation-lancedb==0.45.6
opentelemetry-instrumentation-langchain==0.45.6
opentelemetry-instrumentation-llamaindex==0.45.6
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-instrumentation-marqo==0.45.6
opentelemetry-instrumentation-mcp==0.45.6
opentelemetry-instrumentation-milvus==0.45.6
opentelemetry-instrumentation-mistralai==0.45.6
opentelemetry-instrumentation-ollama==0.45.6
opentelemetry-instrumentation-openai==0.45.6
opentelemetry-instrumentation-openai-agents==0.45.6
opentelemetry-instrumentation-pinecone==0.45.6
opentelemetry-instrumentation-qdrant==0.45.6
opentelemetry-instrumentation-redis==0.55b1
opentelemetry-instrumentation-replicate==0.45.6
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-sagemaker==0.45.6
opentelemetry-instrumentation-sqlalchemy==0.55b1
opentelemetry-instrumentation-threading==0.55b1
opentelemetry-instrumentation-together==0.45.6
opentelemetry-instrumentation-transformers==0.45.6
opentelemetry-instrumentation-urllib3==0.55b1
opentelemetry-instrumentation-vertexai==0.45.6
opentelemetry-instrumentation-watsonx==0.45.6
opentelemetry-instrumentation-weaviate==0.45.6
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.55b1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
posthog==3.25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.1
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==13.9.4
rpds-py==0.23.1
rsa==4.9.1
selenium==4.35.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==3.0.2
starlette==0.47.2
streamlit==1.44.1
syncer==2.0.3
tenacity==9.0.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traceloop-sdk==0.45.6
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==0.24.0
websocket-client==1.8.0
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
